refactor: extracted logic for selecting and setting default avatar into custom hook and used it in signup process.
