+++ src/hooks/useSignup.ts
@@ -1,40 +1,15 @@
 import { createUserWithEmailAndPassword } from "firebase/auth";
 import { setDoc, doc } from "firebase/firestore";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { toast } from "react-toastify";
 import { auth, db } from "../lib/firebase";
 import { upload } from "../lib/upload";
+import { useSelectAvatar } from "./useSelectAvatar";
 
 export const useSignup = () => {
 
     const [isLoading, setIsLoading] = useState(false);
-    const [avatar, setAvatar] = useState<{
-        file: File | null,
-        url: string
-    }>({
-        file: null,
-        url: "/img/avatar-placeholder.png"
-    });
-
-    useEffect(() => {
-        if (!avatar.file) {
-            fetch(avatar.url)
-                .then(res => res.blob())
-                .then(blob => {
-                    const file = new File([blob], "avatar-placeholder.png", { type: "image/png" });
-                    setAvatar(prev => ({ ...prev, file }));
-                });
-        }
-    }, []);
-
-
-    const selectAvatar = (e: any) => {
-        if (e.target.files[0])
-            setAvatar({
-                file: e.target.files[0],
-                url: URL.createObjectURL(e.target.files[0])
-            });
-    };
+    const { avatar, selectAvatar } = useSelectAvatar();
 
     const signUp = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
@@ -51,6 +26,7 @@ export const useSignup = () => {
                 email,
                 id: res.user.uid,
                 avatar: imgUrl,
+                status: "I'm using buzz",
                 blocked: []
             });
             await setDoc(doc(db, "userchats", res.user.uid), {
